Toykeeper conversion + buffered output handling
Output Format
------------

The converters have both been altered so that instead of a log
format they are passed an output_handler object which implements
an output format. This operates on a day-by-day buffer, with
the input converter responsible for building up a day's entry and
passing this to the output handler in the expected pythonic format.

The only implemented output handler at the moment is the JSON
output handler.

As per namespace's request, the Toykeeper log parser reads
line-by-line from files rather than loading them into memory.
The changes to the energymech reader have been minimal, but
are untested due to the lack of test data for that format.

To retain backwards compatibility, leaving the output_handler
as None will cause the converters to merely stash their input
in a python object structure and return it on completion.
This will of course mean greater memory usage than using the
buffered output handler system.

The code for the output handlers is currently in the main
frontend file. It may be worth working out a different location
for this.

Interface
--------

I have tweaked the code in the frontend to some degree. The
changes here are not highly important.