Datestamp fix; STDOUT handling, outline of Sqlite.
The datestamps from Toykeeper are now in the integer format which
corresponds with what the energymech converter users.

The JSON handler now correctly handles stdout, but the way this is
handled should be altered anyway as Sqlite can't write to stdout
sensibly (I think, at any rate).

The Sqlite stuff is very partial and not even remotely tested,
only an outline of what's going to be done. It will not work, and
bits which exist will be wrong.